requirements.txt
streamlit
scikit-learn
pandas
numpy
matplotlib
seaborn
joblib
pdfplumber
PyMuPDF
pytesseract
google-generativeai
